renamed scanner to hook_scanner
