+++ secureli/modules/core/core_services/hook_scanner.py
