+++ tests/modules/core/test_hook_scanner_service.py
@@ -5,7 +5,7 @@ import pytest
 from secureli.modules.shared.abstractions.pre_commit import ExecuteResult
 from secureli.modules.shared.models.scan import OutputParseErrors, ScanMode
 import secureli.modules.shared.models.repository as RepositoryModels
-from secureli.modules.core.core_services.scanner import HooksScannerService
+from secureli.modules.core.core_services.hook_scanner import HooksScannerService
 from pytest_mock import MockerFixture
 
 test_folder_path = Path(".")
